+++ lizard_measure/urls.py
@@ -1,10 +1,12 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
 from django.conf.urls.defaults import include
 from django.conf.urls.defaults import patterns
+from django.conf.urls.defaults import url
 from django.conf import settings
 from django.contrib import admin
 
 from lizard_measure.views import MeasureDetailView
+from lizard_measure.views import HorizontalBarGraphView
 
 
 API_URL_NAME = 'lizard_measure_api_root'
@@ -69,6 +71,19 @@ urlpatterns = patterns(
     #  {},
     #  "lizard_measure.measure_graph"),
     # # Search stuff.
+    # url(r'^homepage_area_search/',
+    #     'lizard_measure.views.waterbody_shapefile_search',
+    #     name="lizard_measure_homepage_area_search"),
+    # # Map clicking, workspaces
+    # (r'^map/', include('lizard_map.urls')),
+    # # Fake help page.
+    # url(r'^$',
+    #     'lizard_measure.views.select_area',
+    #     name="help"
+    #     ),
+    url(r'^bar/$',
+        HorizontalBarGraphView.as_view(),
+        name="lizard_graph_horizontal_bar_graph_view"),
     )